Fixed the camera jitters, which also fixed the jitters when volume control is changed (Issue #6). The order in which everything is updated is very important. Player updates first, then camera, then the Player's sprite, then everything else. Also re-added the turbo.

+++ Clay/GameObject.h
@@ -65,7 +65,7 @@ typedef enum {
 -(void) switchToInactive;
 
 -(void) setPosition:(CGPoint)position;
--(void) setPositionAtX:(int)x Y:(int)y;     //give new x and y position on screen (with cocos2D, both hi and low-res use 320x480 resolution for its points)
+-(void) setPositionAtX:(float)x Y:(float)y;     //give new x and y position on screen (with cocos2D, both hi and low-res use 320x480 resolution for its points)
 
 -(CCSprite*) getCCSprite;
